model sightings controller response by location
We're going to want to give the user the ability to choose how they want to view their sightings. We store sightings by location, with each location tied to a county, region and country. For each location, county, region and country we'll store a lat/lon so that we can get markers on the map without having to talk MapBox